Highly improve makefile
Now
'make' makes lasercake by default.
We don't have to update the Makefile when the list of
   files (or their deps) change.
Recompilation is only done when necessary based on headers
   (admittedly, almost everything includes world.hpp
   which includes almost all the headers, so it's not much
   of an improvement at the moment).

This definitely uses GNU Make features (as opposed to
   generic Make features).

Makefiles are not great, but they're standard
(even OSX uses GNU Make, IIRC, though the BSDs don't by default)
(Windows doesn't, of course, except in mingw or cygwin)
and they work okay.

We may use CMake in the future due to its portability and
   how Qt already uses it (we may use Qt).  We'll see!