Only display the half of tile vertices closer to the viewer.
This made, in pressure_tunnel, the non-GL display code
take 2/3 the time and the GL code just over 1/2 the time.

Unintended consequences:
Transparent tiles block your vision half as much. I kind of like
the new water transparency level.
And, I think, if the view is inside a tile, that tile won't
be displayed.
Shrugs for now!