Tiles: show face iff that direction isn't the same type (ROCK,WATER..).
This looks much better aesthetically! - substances not having
axis aligned walls within them.

I deleted the only-show-tile-faces-that-face-the-player check that
I'd put there.  It had some strangeness in pressure_tunnel that I
could see when combined (&&) with this check - on some sides
the wrong rock-faces were displayed/not.  I did not succeed at
debugging it; dropping it gets rid of the TODO question; and
this new check reduces the number of tile-faces enough that
we get a similarly delightful speed benefit anyway.