Draw objects in order from further to closer.
For translucent objects, this is a useful way to do it.

Opaque objects could all be drawn first if they wanted,
using the OpenGL z-buffer. (Not implemented presently.)
For tile-aligned tiles, the ordering appears to be perfect,
as I planned.
For other objects, it's not as perfect, but okayish.

Right now, lasers cheat and always get drawn in front of
nearly everything, because I was lazy coding it and because
it'd need a bit more thought in order to be truly reliable.

The ordering is Manhattan-distance (rounded / hacked to tile width
and height so that it fits in an array(vector), and is exact multiples of
tiles, and puts lots of tiles together in one bucket.
TODO see if a std::map<fine_scalar, vertices_t> of true Manhattan
distances is just as efficient... probably ultimately that would only
be notably worse than this method if we put thousands of translucent
flying robots in a tight space).