Draw closer faces of tiles after farther faces of tiles.
The CPU time cost in 'drawing' phase of adding this code
for pressure_tunnel appears to be 0-3% (the range is due
to not having more measurement precision).

Acceptable, considering that it lets us make the faces
different colors.

Well... it also makes it *harder* to make, e.g.,
the top face a different color, because it's harder
to find out which face is the 'top' one now! But it
should be quite doable.

If we come up with a different way to arrange for
drawing tile-sides in the right order when it matters,
then I'd be happy to scrap this approach, though I don't see
a superior way for right now (for tiles that are translucent
and have different faces different colors).