double-check that ints are what we expect them to be
At least some C compilers/libraries are known to lie,
so this may be a minor safety net.

It doesn't actually check dynamic behaviour, only
what in-compiler integer math, sizeofs, etc. do,
(which are *supposed* to be the same as runtime.
But hey, if the compiler is already wrong, maybe
it gets that wrong too...)