Re-order worldblock data members
It made stepped_pools init slightly faster,
but mainly it didn't make things worse and
it seems nicer to put big data at the end.

I put a pointer right before the array to make
sure it's aligned at least as much as a pointer is.
I tried out adding __attribute__((aligned(n))) and
n=16 or n=32 didn't really make a difference and
n=64 made things a tad slower - well, it does make
the struct bigger in this order, and we're probably
not doing anything that *uses* such alignedness.

I then tried making it a pointer to a new'd array,
and that made the simulation and initing about 5-8% slower.
(That may be an acceptable cost to pay later if it
brings benefits we can't get a better way.)