Address a bbox_collision_detector TODO-comment
"TODO address the fact that this could put bits higher than
the appropriate amount if CoordinateBits isn't the number of
bits of the type"

If CoordinateBits is smaller than the size of type Coordinate,
it's intentional that we nevertheless return 0 as if Coordinate
were no bigger than CoordinateBits.

"CoordinateBits should usually be 32 or 64. I don't know if
it works for other values." is still true; it's quite untested
nor thought about thoroughly.

(We could revert this TODO-addressing someday if it actually
turns out to be more useful to do it the other way for
non-32-or-64 CoordinateBits.)