num_low_bits_ignored_ -> num_low_bits_ (nix "ignored" monikers)
The number of low bits has a specific meaning.  Sure,
that meaning is implemented in part by setting that many
low bits in interleaved_bits_ to zero and/or not doing math
that depends on those bits,
but that's just not very important compared to its meaning.
(see comments above struct zbox for a description of the meaning.)