Use stack not set for bbox_collision_detector searches.
This fixes a latent bug where custom_walker's std::set
given the default bbox_less_than would only ever have one member
(since the comparison function always returned false,
making them equal in the eyes of 'set').
We decided it was not very useful to have set-ordering
anyway (it was a TODO anyway, rethinking the use of std::set).

This change means currently we don't ask the comparison
function to distinguish how depth-first or breadth-first
our search is (it's always very depth-first).  We
didn't think of a use for allowing ordering other than
depth-first, and if we do, we can implement something
correspondingly.