Make code more Clang+libc++ compatible
(In libc++, std::function's copy-constructor requires the arguments
of the function type to be complete types.  So don't use that
copy-constructor in world.hpp.)

I'm not sure whether it's "required" but I think it makes nicer
code anyway, in this case.

(see my SO question; skimming N1836 didn't enlighten me.
http://stackoverflow.com/questions/10730682/does-stdfunctions-copy-constructor-require-the-template-types-argument-types
)

(Clang 3.0 still segfaults trying to compile our lambdas, but
this fix is something.)