Put most bbox_collision_detector code into new .cpp
This required various other minor code changes; pretty much
the semantics are the same but the namespaces and similar
are different.

Using a .cpp is better code design.  It means that changes to most
bbox_collision_detector code won't require recompiling most of
the entire project (it's included in world.hpp and thus most places).

This change:
sped up(?!) simulation speed by about 6%.  I'm guessing this is because
of not inlining huge functions that obviously shouldn't be, and instead
being able to share them between files/usages, thus sharing more code
in the processor's instruction-cache (but I could be wrong. Anyway,
such a speedup is nice!).

slowed(?!) compile time by about 6%.  I was hoping it'd be faster
because not so much code repeated in headers.  Well, there *is* an
extra .cpp now.

I also found a better (more reliable & shorter code) hack than before
for ostream operator<< of templates,
  friend inline std::ostream& operator<<(std::ostream& os, cls const& c) {
    ...
  }
described at http://stackoverflow.com/a/10652104
I consider it a hack when the given operator<< does not need
to be a friend (nor inline) for any other reasons!
The previous hack broke with this refactor (not sure why), so I
found and switched to this.