get_un_bounds_checked_int() -> get_primitive_int() etc
These functions are mainly used in order to use a library
that doesn't get along with the bounds-checked types.
If we use Boost.Units there will be a similar issue
which I suspect may cause problems in most of the same
situations.  So I chose a new, shorter name to describe
this purpose of giving a number to somewhere that needs
primitive int types.

(Downside: It's not as long and ugly a name as
get_un_bounds_checked_*.  But being a function call at all
should discourage people writing unnecessary un-bounds-checking
of things!)