Test & support modulo bboxes for collision_detector
This
(1) does not slow down Lasercake
(2) I always intended this to be the case, as it makes mapping to
 a signed coordinate type in the rest of the program work properly
 even when bounding-boxes straddle the 0 of some dimension.

There's also an assert() and
  if(base_box_size == 0) {..} else
added to make sure insert_zboxes_from_bbox() doesn't try to make
double-size zboxes for things that are going to have a full-space zbox
already.  (The logic there might have to be a bit more complicated
if we enhance insert_zboxes_from_bbox() later in various ways to find
more-minimal boxes.  But it's fine presently.)