fix minor potential behavior change & overflow
that was introduced by implicit bbox conversion patch.

The "- 1" in "bb.size[X] - 1" etc. fixes the behavior change and
associated overflow.  Then, the "- 1" in "(1LL << 32) - 1" just makes
it more conservative.  None of these changes the behavior in practice
because none of our bounding-boxes in our lasercake worlds
are anywhere near that big.

(In the next patch, I'll change size to size_minus_one in places,
because that's almost the only correct way to refer to size!)