Make collision_detector::bounding_box an abstract type.
This makes it easier to change its exact representation,
should that become important.  It makes it easier to
access any of min, max or size in ways that look parallel
(which is good because collision_detector::bounding_box
is part of the bbox_collision_detector interface).

Its effect on performance is somewhere from 0 to -0.5%
(not sure); not worth worrying about for a code cleanup
that doesn't change the amount of computation involved.