Use Boost.Chrono instead of POSIX/Linux clock_gettime() librt
Also allow disabling it entirely with CMake option TIMING=OFF.

This is intended to aid portability.

It also makes the code a bit nicer.

(The code could be nicer still if we used the
chrono::duration type in more places in main.cpp
rather than the int64_t microseconds_t.)