Move tile_coordinate<->fine_scalar conversions to tiles.hpp.
More specific places are good (and this piece of code is going
to get bigger soon).

It wasn't necessarily part of tiles.hpp before to have a concept
of fine_scalar space at all.  But I think it's okay for it (and
only this section of it, at that!) to have such a concept.