Fix beam_first_contact_finder virtual method naming.
I don't notice any practical effect from making this change

(I did check that I'm now overriding by using C++11 'override'
contextual-keyword, which I'm not leaving in the commit because
not enough compilers support it and the benefit of leaving it
there seems small.  We don't use virtuals/overriding enough to
make too many of these mistakes.)

It turns out that the laser irregularities I was noticing
mostly had to do with other uncommitted changes I had [and have
not committed as of this commit].  I think the lasers are fine
before and after this commit, and after those will-be-committed-next
changes too - it was probably some intermediate state screwing
things up.