CMakeLists: use cmake sequences not strings for flag computation
- why not use something more semantic to the language,
to as far as it lets us.

It was also necessary for (not planning to commit the
following) the purpose of calling GCC manually use
pre-compiled headers, which I gave up on because
(the way I tried it at least) it didn't always decrease
compilation time (sometimes increased!) and never
decreased it by much, and caused a few other
complications with compiler warnings and with
recompilation-checking.