Change polygon_collision_detection to use faux optionals
Optionals are more semantic for its purpose than the
pair<bool, rational>, but they were a tad slower, so I made a
faux_optional<> with the same interface as boost::optional<>
that is as fast as pair<bool, rational> in this case
(with this optimizer - GCC 4.7 -O3).

(This speed kinda makes sense in that boost::optional<> has to obscure
its value storage memory to the compiler a bit - have it be a
different type - so that that value can be non-constructed sometimes.
non_normalized_rational is just a boring pair of integers that costs
nothing to construct, so boost::optional doesn't save that cost as
it might with another member type.)