rename a fine_scalar->tile function
convert_to_smallest_superset_at_tile_resolution
->
get_tile_bbox_containing_all_tiles_intersecting_fine_bbox

because the latter is clearer about what we need in the one case
where we need it, in the (upcoming) case where tiles overlap
at their edges in fine_scalar space.

Note that this will (in new not-yet-committed tile semantic)
make alternated applications of
  convert_to_fine_units(tile_bounding_box)
and
  get_tile_bbox_containing_all_tiles_intersecting_fine_bbox(bounding_box)
gradually increase the size of the box every time
you repeat it.  But that's an intrinsic cost we pay
for making tiles overlap (i.e. a semantic we have to
be aware of while writing simulation code).  See
soon-upcoming comment for why *not* doing so is worse.