Abstraction for ::bounding_box and related changes
This brings ::bounding_box's interface more in line with
collision_detector::bounding_box's.
It also caught a bug which I fixed - rendering lasers
doesn't actually use a proper min and max, so I added a
spanning_two_points() bounding-box-creator.

On gcc 4.7 -O3, this seems to cost 5% performance :-(
(pressure_tunnel)
but is probably worth the abstraction, consistency,
bug-catching-ness... On clang 3.1 it doesn't cost
so much performance.

TODO: similar abstraction for tile_bounding_box? Not
as much code uses that, so it's not *as* important
for coding reasons.