Change tile_width and tile_height.
The new values make
a fine_scalar unit = 1 millimetre
a tile_width = 2 metre
a tile_height = 0.4 metre

- usable proportions once we put it in the game interface.

(We already decided/guessed that, for what we're planning to use
tiles for, 2 metres is more practical than 1 metre.  E.g. putting
a railroad track along/within a tile.)

This has a performance cost in the draw code (5-20%),
but it looks like it's all in the bbox_collision_detector iteration,
so this just adds importance to our hopes to take the tiles out of the
fine_scalar ztree (issue:
https://github.com/Lasercake/Lasercake/issues/25
)