Add ability to use a local copy of Boost.
Motivating reasons are described in a footnote in
generate-boostbcp.py

This does increase my size of the lasercake binary itself
from about 1 MB to about 1.5 MB.  This overhead is commensurate to
or less than the size of my .so library files for the ones we link to.
Also, 1.5 MB is still pretty small for a serious GUI-program binary.