Don't draw back (from viewer's point of view) surfaces of tiles.
This makes the graphics card happier (half as much to draw).
It looks different for translucent tile types - they are twice
as transparent at before - but of course we choose opacity accordingly.

Also, it means that we never draw two neighboring tiles, one of which
is translucent, with a face in the exact same OpenGL-place.  This means
we don't have to think about which one is on top!