Save more init time
This saves another 30 ms init time.

We decided to make worldblock data public:
they are a hacky performance thing anyway, and the abstraction
needed to make them a hacky, encapsulated performance thing
would obscure as much as it revealed.  The simple act of accessing
a 'worldblock' indicates you're doing something funny already.

So, use 'git grep' or such to find out who's mucking with the members.