Use our deque in tile_physics.
This didn't make a noticable performance impact in
pressure_tunnel -lne715.  It exercises the deque code,
and might be good for something (the profile of tile_physics code
for pressure_tunnel looked uglier previously [although it might
have been one of the other intervening changes that changed
that -- I'm not spending the time now to run both in
valgrind --tool=callgrind to check]).