Speed up prepare_tile() by about 10%
(...simply re-ordering the code made it measurably faster. Huh.
I wonder if it was something to do with cache/prefetch, though when I
just before this tried adding
__builtin_prefetch(coll.quads.vertices + coll.quads.count, 1);
it didn't help.
Callgrind had told me that there were a LOT of write cache-misses
at both L1 and LL, though I'm not sure of the best way to interpret its
data.)