fix build when boost::uint_t<32|64>::fast is neither uint32_t nor uint64_t
e.g. on my OS X 10.6 with gcc-mp-4.7 - there is "int", "long" and
"long long" and equal sizes are not the same type, and boost and
stdint.h happen to make different choices from each other.

This change also allows you to call these functions with a greater
variety of types, appropriately checked for correctness, which may
be useful.