Change GL scale to fine_scalar scale.
This means that tile-height locations are exact floating-point
locations now, yay!  Also it just makes more sense.  And we still
have quite a bit of exact range in 'float', and quite a bit more
as it floats before it would mix up the order of two tiles,
so I think we're fine on that count.

This has no noticable effect on prep or GL speed.