Fix patricia_trie to (hypothetically) work with big-ints.
Hypothetically = I haven't tested with any actual bigints,
and foresee no need for us using this, but it was silly
to not allow that when it was so close to being true already.
Also, "assert(shared_size_exponent < coordinate_bits);" was
wrong. (it would have been trivially but uselessly correct
with <=).