use std::boolalpha on cerr: attempt to get confused less
When we write a debug message wrong and something gets implicitly
converted to bool, outputting it as 0 or 1 can confuse it with
an int.  This happened a few times.  I haven't checked whether
anything unsets boolalpha during the run of Lasercake, though.
(Our code doesn't unset it, but perhaps a library might.)