Add tile-iteration visitor that doesn't use bbox_collision_detector.
And use it in the display code.  This incidentally speeds up(!)
`spiky -ne35` gl-prep time by around 10% compared to what it was
with the similar code before.  I suspect it's the more thorough
look_here() tribool management that's responsible for the speed.
It could also be a bit from not using the bbox_collision_detector
of worldblocks in favor of the new trie, though that has so few
items either way that I doubt that mattered much.