Add a way to make more-deterministic pseudo-random scenarios.
And add a version of spiky that uses it.  I'm getting
spiky2 250ms init
spiky3 270ms init
which is not a bad price to pay, although if we can eliminate
some of the other costs of worldblock initing then the RNG
initialization will be a higher cost to pay.

The taus88 generator was plenty fast but generated
obviously-not-random-enough worlds.  The Mersenne twister variations
were somewhat slower to init.  ranlux seems to be a fine
in-between point.  Boost docs say its speed per random number
is quite low, but for these memo purposes each instance of the RNG
is not making that many numbers, so that faster init time than
mersenne-twister beats mersenne-twister in speed here.

There are other ways to do it that I haven't tried
(e.g. if you need very few random bits, you could MD4 hash the
arguments to get those pseudorandom bits; I didn't have a MD4
implementation on hand to test that's speed quickly).
This is fine for now.