Make simple_hills use newer infrastructure.
This only decreases its worldgen runtime by about 10%.
Halving max_simple_hill_width makes the runtime about 1/4,
which means that the cost is linear in that:
i.e., the cost to look up all those memos in a hashtable for
every x and y, since the memo shares most of the actual computations
of get_hill.  It's interesting that there wasn't a significant
improvement from changing from tile-spec to column-spec;
maybe the optimizer was good enough to share some of the lookup work
in the previous version (worldgen_from_tilespec_t makes it plausible
the optimizer might do something like that).