Make the rendering thread semi-interruptable.
Now the rendering thread checks an atomic flag frequently
and if/when it's set, returns early rather than finishing the
potentially large number of OpenGL calls it was queued up to do.
This makes quitting with Esc quick, without risking the wrath of
interrupting a GL implementation mid-call nor making it hard to
deallocate the GL buffers we allocate.