make WARNING_FLAGS only use flags that the compiler recognizes,
and add Clang's equivalent of GCC's fmax-errors, and
reduce the number to a limit somewhat more useful with
huge template error messages.

In order to get Clang not to warn about -ferror-limit every time,
we had to add -Werror to the checks to that Clang's warning
about not recognizing -fmax-errors could be upgraded to an error.
LOL.  Also check_cxx_compiler_flag cannot handle flags that have
spaces in them, because it splits on spaces...