Add (Clang-only) warning flag for global constructors & destructors.
They are inefficient, and racy.  They bloat startup/shutdown time and
binary size.  Destructors race between one thread calling exit()
and destroying global objects, and other threads using them.
Constructors in separate translation units have arbitrary
initialization order.  Neither of these is something
we can't deal with moderately easily, but it's nice not to have any
and we can do it!  Also, the work in doing it possibly helps things
in other ways.

Setting this warning is the first step to finding and eliminating them.