constexpr another global.
Now the only global-constructors left are the tests,
and boost test framework with the auto test cases works that way.
I'll be able to fix that too, and it might even help fix the
way my clang's build is looping when it tries to run the tests...
in a bit, though.