Update Clang requirement
I built a Clang 3.1 to test with; it gave a fixable error, followed
by crashing on lots of code, which sounds familiar from last time I
tried that version of Clang on Lasercake (it worked at some code
states...).  Probably not worth supporting it.  Yell if you're stuck
on a version of Xcode with a Clang snapshot that can't quite build this
and if you can't use compilers from Macports?  Let me know if this is
a common situation.