Add release-build script for reproducibly building releases
It likely doesn't meet the standards of deterministic builds
( http://gitian.org/ , https://wiki.debian.org/ReproducibleBuilds )
yet, but it makes it much easier to make releases and to
be confident that I'll be able to know with high confidence,
even years later, exactly how the releases were made.

This script currently supports building for Linux and Windows.